+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Warn about function body not being used, because it already has external
   implementations for all targets.
+- It's now possible to compile a project with external functions that are not
+  supported by the compilation target as long as those are not actually used.
 
 ### Language Server Changes